refactor: Move Attendee detailed enum types from entity package to entity.details package
